[parser]: extract editor-parser and text-parser packages from parser
