+++ packages/editor-parser/package.json
@@ -1,0 +1,4 @@
+{
+    "name": "@math-blocks/editor-parser",
+    "version": "0.0.1"
+}
