+++ packages/editor-parser/src/__tests__/editor-parser.test.ts
@@ -2,7 +2,7 @@ import * as Editor from "@math-blocks/editor";
 
 import parser from "../editor-parser";
 import * as Lexer from "../editor-lexer";
-import * as LexUtil from "../lexer-ast";
+import * as LexUtil from "../test-util";
 
 type Loc = {};
