+++ packages/editor-parser/src/editor-lexer.ts
