+++ packages/editor-parser/src/index.ts
@@ -1,0 +1,1 @@
+export {parse} from "./editor-parser";
