+++ packages/editor/src/index.ts
@@ -1,6 +1,8 @@
-import reducer, {State} from "./editor-reducer";
+import reducer from "./editor-reducer";
 import print from "./editor-printer";
 import * as Util from "./util";
+
+export {reducer, print, Util};
 export * from "./editor-ast";
+export * from "./editor-reducer";
 export {isEqual} from "./util"; // TODO: dedupe methods in editor and util
-export {State, reducer, print, Util};
