+++ packages/step-checker/src/__tests__/integer-checker.test.ts
@@ -1,5 +1,5 @@
 import {serializer} from "@math-blocks/semantic";
-import {parse} from "@math-blocks/parser";
+import {parse} from "@math-blocks/text-parser";
 
 import StepChecker from "../step-checker";
 import {Result} from "../types";
