+++ packages/text-parser/package.json
@@ -1,0 +1,4 @@
+{
+    "name": "@math-blocks/text-parser",
+    "version": "0.0.1"
+}
