+++ packages/text-parser/src/__tests__/text-lexer.test.ts
