+++ packages/text-parser/src/__tests__/text-parser.test.ts
