+++ packages/text-parser/src/text-lexer.ts
