+++ packages/text-parser/src/text-parser.ts
@@ -1,6 +1,6 @@
 import * as Semantic from "@math-blocks/semantic";
+import * as Parser from "@math-blocks/parser";
 
-import * as Parser from "./parser";
 import {lex, Token} from "./text-lexer";
 
 // TODO: fill out this list
