+++ webpack.config.js
@@ -51,6 +51,14 @@ module.exports = {
                 __dirname,
                 "./packages/parser/src",
             ),
+            "@math-blocks/editor-parser": path.join(
+                __dirname,
+                "./packages/editor-parser/src",
+            ),
+            "@math-blocks/text-parser": path.join(
+                __dirname,
+                "./packages/text-parser/src",
+            ),
             "@math-blocks/typesetter": path.join(
                 __dirname,
                 "./packages/typesetter/src",
